requirements.txt
asgiref==3.2.7
Django==3.0.4
django-crispy-forms==1.9.0
django-dynamic-formsets==0.0.8
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
